MOD: FieldAndField & FieldAndValue upper class
