+++ protoframework-sql/src/test/java/org/protoframework/orm/sql/TestArithmeticExpr.java
@@ -23,7 +23,7 @@ public class TestArithmeticExpr {
 
     {
       // +
-      expr = Expressions.FieldAndValue.add("a", 1);
+      expr = FieldAndValue.add("a", 1);
       System.out.println(expr);
       assertTrue(expr.getLeft() instanceof Column);
       assertTrue(expr.getRight() instanceof Value);
@@ -37,19 +37,19 @@ public class TestArithmeticExpr {
     }
     {
       // -
-      expr = Expressions.FieldAndValue.subtract("a", 1);
+      expr = FieldAndValue.subtract("a", 1);
       System.out.println(expr);
       assertEquals(ArithmeticOp.SUBTRACT, expr.getOp());
       assertEquals("a-?", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("a-1", expr.toSolidSql(new StringBuilder()).toString());
-      expr = Expressions.FieldAndValue.subtract(1, "a");
+      expr = FieldAndValue.subtract(1, "a");
       System.out.println(expr);
       assertEquals("?-a", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("1-a", expr.toSolidSql(new StringBuilder()).toString());
     }
     {
       // *
-      expr = Expressions.FieldAndValue.multiply("a", 1);
+      expr = FieldAndValue.multiply("a", 1);
       System.out.println(expr);
       assertEquals(ArithmeticOp.MULTIPLY, expr.getOp());
       assertEquals("a*?", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -57,36 +57,36 @@ public class TestArithmeticExpr {
     }
     {
       // /
-      expr = Expressions.FieldAndValue.divide("a", 1);
+      expr = FieldAndValue.divide("a", 1);
       System.out.println(expr);
       assertEquals(ArithmeticOp.DIVIDE, expr.getOp());
       assertEquals("a/?", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("a/1", expr.toSolidSql(new StringBuilder()).toString());
-      expr = Expressions.FieldAndValue.divide(1, "a");
+      expr = FieldAndValue.divide(1, "a");
       System.out.println(expr);
       assertEquals("?/a", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("1/a", expr.toSolidSql(new StringBuilder()).toString());
     }
     {
       // div
-      expr = Expressions.FieldAndValue.divRound("a", 1);
+      expr = FieldAndValue.divRound("a", 1);
       System.out.println(expr);
       assertEquals(ArithmeticOp.DIV_ROUND, expr.getOp());
       assertEquals("a DIV ?", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("a DIV 1", expr.toSolidSql(new StringBuilder()).toString());
-      expr = Expressions.FieldAndValue.divRound(1, "a");
+      expr = FieldAndValue.divRound(1, "a");
       System.out.println(expr);
       assertEquals("? DIV a", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("1 DIV a", expr.toSolidSql(new StringBuilder()).toString());
     }
     {
       // mod
-      expr = Expressions.FieldAndValue.mod("a", 1);
+      expr = FieldAndValue.mod("a", 1);
       System.out.println(expr);
       assertEquals(ArithmeticOp.MOD, expr.getOp());
       assertEquals("a MOD ?", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("a MOD 1", expr.toSolidSql(new StringBuilder()).toString());
-      expr = Expressions.FieldAndValue.mod(1, "a");
+      expr = FieldAndValue.mod(1, "a");
       System.out.println(expr);
       assertEquals("? MOD a", expr.toSqlTemplate(new StringBuilder()).toString());
       assertEquals("1 MOD a", expr.toSolidSql(new StringBuilder()).toString());
@@ -100,7 +100,7 @@ public class TestArithmeticExpr {
 
     {
       // +
-      expr = Expressions.FieldAndField.add("a", "b");
+      expr = FieldAndField.add("a", "b");
       System.out.println(expr);
       assertTrue(expr.getLeft() instanceof Column);
       assertTrue(expr.getRight() instanceof Column);
@@ -112,7 +112,7 @@ public class TestArithmeticExpr {
     }
     {
       // -
-      expr = Expressions.FieldAndField.subtract("a", "b");
+      expr = FieldAndField.subtract("a", "b");
       System.out.println(expr);
       assertEquals(ArithmeticOp.SUBTRACT, expr.getOp());
       assertEquals("a-b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -120,7 +120,7 @@ public class TestArithmeticExpr {
     }
     {
       // *
-      expr = Expressions.FieldAndField.multiply("a", "b");
+      expr = FieldAndField.multiply("a", "b");
       System.out.println(expr);
       assertEquals(ArithmeticOp.MULTIPLY, expr.getOp());
       assertEquals("a*b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -128,7 +128,7 @@ public class TestArithmeticExpr {
     }
     {
       // /
-      expr = Expressions.FieldAndField.divide("a", "b");
+      expr = FieldAndField.divide("a", "b");
       System.out.println(expr);
       assertEquals(ArithmeticOp.DIVIDE, expr.getOp());
       assertEquals("a/b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -136,7 +136,7 @@ public class TestArithmeticExpr {
     }
     {
       // div
-      expr = Expressions.FieldAndField.divRound("a", "b");
+      expr = FieldAndField.divRound("a", "b");
       System.out.println(expr);
       assertEquals(ArithmeticOp.DIV_ROUND, expr.getOp());
       assertEquals("a DIV b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -144,7 +144,7 @@ public class TestArithmeticExpr {
     }
     {
       // mod
-      expr = Expressions.FieldAndField.mod("a", "b");
+      expr = FieldAndField.mod("a", "b");
       System.out.println(expr);
       assertEquals(ArithmeticOp.MOD, expr.getOp());
       assertEquals("a MOD b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -155,7 +155,7 @@ public class TestArithmeticExpr {
   @Test
   public void testEmbedding() {
     ArithmeticExpr expr = ArithmeticExpr.add(
-        Expressions.FieldAndValue.add("a", 1), Expressions.FieldAndField.add("a", "b"));
+        FieldAndValue.add("a", 1), FieldAndField.add("a", "b"));
     System.out.println(expr);
     assertEquals("a+?+a+b", expr.toSqlTemplate(new StringBuilder()).toString());
     assertEquals("a+1+a+b", expr.toSolidSql(new StringBuilder()).toString());
@@ -165,12 +165,12 @@ public class TestArithmeticExpr {
     assertEquals(1, sqlValues.get(0).getValue());
 
     expr = ArithmeticExpr.add(
-        Expressions.FieldAndField.add("a", "b"), Expressions.FieldAndField.multiply("c", "d"));
+        FieldAndField.add("a", "b"), FieldAndField.multiply("c", "d"));
     System.out.println(expr);
     assertEquals("a+b+(c*d)", expr.toSqlTemplate(new StringBuilder()).toString());
 
     expr = ArithmeticExpr.add(
-        Expressions.FieldAndField.eq("a", "b"), Expressions.FieldAndField.and("c", "d"));
+        FieldAndField.eq("a", "b"), FieldAndField.and("c", "d"));
     System.out.println(expr);
     assertEquals("(a=b)+(c AND d)", expr.toSqlTemplate(new StringBuilder()).toString());
   }
