+++ protoframework-sql/src/test/java/org/protoframework/orm/sql/TestLogicalExpr.java
@@ -24,7 +24,7 @@ public class TestLogicalExpr {
 
     {
       // and
-      expr = Expressions.FieldAndField.and("a", "b");
+      expr = FieldAndField.and("a", "b");
       System.out.println(expr);
       assertTrue(expr.getLeft() instanceof Column);
       assertTrue(expr.getRight() instanceof Column);
@@ -36,7 +36,7 @@ public class TestLogicalExpr {
     }
     {
       // or
-      expr = Expressions.FieldAndField.or("a", "b");
+      expr = FieldAndField.or("a", "b");
       System.out.println(expr);
       assertEquals(LogicalOp.OR, expr.getOp());
       assertEquals("a OR b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -44,7 +44,7 @@ public class TestLogicalExpr {
     }
     {
       // xor
-      expr = Expressions.FieldAndField.xor("a", "b");
+      expr = FieldAndField.xor("a", "b");
       System.out.println(expr);
       assertEquals(LogicalOp.XOR, expr.getOp());
       assertEquals("a XOR b", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -52,7 +52,7 @@ public class TestLogicalExpr {
     }
     {
       // not
-      expr = Expressions.FieldAndField.not("a");
+      expr = FieldAndField.not("a");
       System.out.println(expr);
       assertEquals(LogicalOp.NOT, expr.getOp());
       assertEquals("NOT a", expr.toSqlTemplate(new StringBuilder()).toString());
@@ -65,7 +65,7 @@ public class TestLogicalExpr {
   @Test
   public void testEmbedding() {
     IExpression expr = LogicalExpr.and(
-        Expressions.FieldAndValue.add("a", 1), Expressions.FieldAndField.add("a", "b"));
+        FieldAndValue.add("a", 1), FieldAndField.add("a", "b"));
     System.out.println(expr);
     assertEquals("(a+?) AND (a+b)", expr.toSqlTemplate(new StringBuilder()).toString());
     assertEquals("(a+1) AND (a+b)", expr.toSolidSql(new StringBuilder()).toString());
@@ -75,21 +75,21 @@ public class TestLogicalExpr {
     assertEquals(1, sqlValues.get(0).getValue());
 
     expr = LogicalExpr.and(
-        Expressions.FieldAndField.and("a", "b"), Expressions.FieldAndField.and("c", "d"));
+        FieldAndField.and("a", "b"), FieldAndField.and("c", "d"));
     System.out.println(expr);
     assertEquals("a AND b AND c AND d", expr.toSqlTemplate(new StringBuilder()).toString());
 
     expr = LogicalExpr.and(
-        Expressions.FieldAndField.or("a", "b"), Expressions.FieldAndField.and("c", "d"));
+        FieldAndField.or("a", "b"), FieldAndField.and("c", "d"));
     System.out.println(expr);
     assertEquals("(a OR b) AND c AND d", expr.toSqlTemplate(new StringBuilder()).toString());
 
     expr = LogicalExpr.and(
-        Expressions.FieldAndField.gt("a", "b"), Expressions.FieldAndField.and("c", "d"));
+        FieldAndField.gt("a", "b"), FieldAndField.and("c", "d"));
     System.out.println(expr);
     assertEquals("a>b AND c AND d", expr.toSqlTemplate(new StringBuilder()).toString());
 
-    expr = Expressions.FieldAndValue.gt("a", 1);
+    expr = FieldAndValue.gt("a", 1);
     expr = expr.and(Column.of("b"));
     System.out.println(expr);
     assertEquals("a>? AND b", expr.toSqlTemplate(new StringBuilder()).toString());
